refactor: centralize storage key helpers
